# backend/gestion-evento-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Eventos",
        "description": "Documentación de la API para el proyecto de gestión de eventos",
        "contact": {
            "email": "soporte@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios",
                "description": "Devuelve todos los usuarios disponibles",
                "operationId": "c8cfdac38d2a424a67cd765f37e8b85d",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuario"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "description": "Crea un usuario con nombre y opcionalmente una foto",
                "operationId": "594d7281b042eb7156b0d6594125840f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "Administrador@gmail.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "administrador123"
                                    },
                                    "escuela_id": {
                                        "description": "Escuela_id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "role_id": {
                                        "description": "Role_id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "nombres": {
                                        "description": "Nombres",
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "apellidos": {
                                        "description": "Apellidos",
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "tipo_documento": {
                                        "description": "Tipo_documento",
                                        "type": "string",
                                        "example": "DNI"
                                    },
                                    "numero_documento": {
                                        "description": "numero_documento",
                                        "type": "string",
                                        "example": "1234567"
                                    },
                                    "telefono": {
                                        "description": "Telefono",
                                        "type": "string",
                                        "example": "1234567"
                                    },
                                    "direccion": {
                                        "description": "Direccion",
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "correo_electronico": {
                                        "description": "Correo_electronico",
                                        "type": "string",
                                        "example": "1234567@gmail.com"
                                    },
                                    "fecha_nacimiento": {
                                        "description": "Fecha_nacimiento",
                                        "type": "string",
                                        "example": "2000-07-10"
                                    },
                                    "codigo_universitario": {
                                        "description": "Codigo_universitario",
                                        "type": "string",
                                        "example": "1234567"
                                    },
                                    "carrera": {
                                        "description": "Carrera",
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "ciclo": {
                                        "description": "Ciclo",
                                        "type": "string",
                                        "example": "1234567"
                                    },
                                    "biografia": {
                                        "description": "Biografia",
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "ponencia_id": {
                                        "description": "Ponencia_id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "especialidad": {
                                        "description": "Especialidad",
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "foto_perfil": {
                                        "description": "Imagen del usuario (archivo opcional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El campo nombre es obligatorio."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Alumno": {
                "title": "Alumno",
                "description": "Datos de un alumno asociado al usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "userId": {
                        "type": "integer",
                        "example": 11
                    },
                    "codigoUniversitario": {
                        "type": "string",
                        "example": "20201234",
                        "nullable": true
                    },
                    "carrera": {
                        "type": "string",
                        "example": "Ingeniería de Sistemas"
                    },
                    "ciclo": {
                        "type": "string",
                        "example": "VII"
                    }
                },
                "type": "object"
            },
...
            "Persona": {
                "title": "Persona",
                "description": "Datos de la persona asociada a un usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "userId": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "nombres": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Pérez García"
                    },
                    "tipoDocumento": {
                        "type": "string",
                        "example": "DNI"
                    },
                    "numeroDocumento": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+51 987654321"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Siempre Viva 123"
                    },
                    "correoElectronico": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@ejemplo.com"
                    },
                    "fotoPerfil": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/usuarios/juan.png",
                        "nullable": true
                    },
                    "fechaNacimiento": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-20"
                    }
                },
                "type": "object"
            },
            "Ponente": {
                "title": "Ponente",
                "description": "Datos del ponente asociado a un usuario y una ponencia",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "biografia": {
                        "type": "string",
                        "example": "Ingeniero con 10 años de experiencia en inteligencia artificial.",
                        "nullable": true
                    },
                    "userId": {
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "ponenciaId": {
                        "description": "Identificador de la ponencia asignada",
                        "type": "integer",
                        "example": 45
                    }
                },
                "type": "object"
            },
            "Role": {
                "title": "Role",
                "description": "Rol del sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Administrador"
                    },
                    "foto": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/roles/admin.png"
                    }
                },
                "type": "object"
            },
            "Usuario": {
                "title": "Usuario",
                "description": "Entidad Usuario del sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@ejemplo.com"
                    },
                    "passwordHash": {
                        "type": "string",
                        "example": "$2y$10$AbCdEf123456..."
                    },
                    "escuelaId": {
                        "type": "integer",
                        "example": 5
                    },
                    "roleId": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "emailVerifiedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-04T10:15:30Z",
                        "nullable": true
                    },
                    "rememberToken": {
                        "type": "string",
                        "example": "a1b2c3d4e5f6g7",
                        "nullable": true
                    },
                    "persona": {
                        "$ref": "#/components/schemas/Persona"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    },
                    "ponente": {
                        "$ref": "#/components/schemas/Ponente"
                    },
                    "alumno": {
                        "$ref": "#/components/schemas/Alumno"
                    },
                    "jurado": {
                        "$ref": "#/components/schemas/Jurado"
                    },
                    "foto": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/usuarios/admin.png"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token de acceso obtenido al iniciar sesión con Laravel Passport",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}